cdp-langchain/examples/chatbot-python/pyproject.toml
[tool.poetry]
name = "chatbot-python"
version = "0.0.1"
description = "CDP AgentKit Example Chatbot"
authors = ["John Peterson <john.peterson@coinbase.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
cdp-langchain = { path = "../../python", develop = true }
bip-utils = "^2.9.3"
py-sr25519-bindings = "^0.2.1"
py-dotenv = "^0.1"
huggingface-hub = "^0.28.1"
transformers = "^4.48.2"
langchain-community = "^0.3.17"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
